♻ extract selectNextCard
