+++ elements/reigns/src/features/game/gameSlice.test.ts
@@ -2,7 +2,6 @@ import { GamePhase } from "../../constants";
 import {
   answerNo,
   answerYes,
-  cardsDistributedByWeight,
   gameReducer,
   initializeGame,
   initialState,
@@ -48,18 +47,6 @@ describe("gameReducer", () => {
         initializeGame.fulfilled(getState([]).definition, "", "")
       );
       expect(result.phase).toBe(GamePhase.ERROR);
-    });
-  });
-
-  describe("cardsDistributedByWeight", () => {
-    it("distributes cards by weight", () => {
-      const result = cardsDistributedByWeight([
-        { card: "card1", weight: 1 } as Card,
-        { card: "card2", weight: 5 } as Card,
-      ]);
-
-      expect(result.filter((p) => p.card === "card1").length).toBe(1);
-      expect(result.filter((p) => p.card === "card2").length).toBe(5);
     });
   });
 });
